Use split-file for tests
