--- clang/test/Modules/Inputs/SafeBuffers/safe_buffers_test.modulemap
@@ -1,15 +1,0 @@
-module safe_buffers_test_base {
-  header "base.h"
-}
-
-module safe_buffers_test_module {
- textual header "test_module.h"
-}
-
-module safe_buffers_test_optout {
-  explicit module test_sub1 {  header "test_sub1.h" }
-  explicit module test_sub2 {  textual header "test_sub2.h" }
-  use safe_buffers_test_base
-  use safe_buffers_test_module
-}
-
