--- clang/test/Modules/Inputs/SafeBuffers/test_module.h
@@ -1,7 +1,0 @@
-#ifdef __cplusplus
-int test_module(int *p) {
-  int x = p[5];
-  int y = p[5];
-  return x + y;
-}
-#endif
