--- clang/test/Modules/Inputs/SafeBuffers/test_sub1.h
@@ -1,20 +1,0 @@
-#include "base.h"
-
-#ifdef __cplusplus
-int sub1(int *p) {
-  int x = p[5];
-#pragma clang unsafe_buffer_usage begin  
-  int y = p[5];
-#pragma clang unsafe_buffer_usage end
-  return x + y;
-}
-
-template <typename T>
-T sub1_T(T *p) {
-  T x = p[5];
-#pragma clang unsafe_buffer_usage begin  
-  T y = p[5];
-#pragma clang unsafe_buffer_usage end
-  return x + y;
-}
-#endif
